pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scaleio"
version = "0.1.0"
description = "AI Image Upscaler using Real-ESRGAN"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Scaleio Team", email = "team@scaleio.ai"}
]
dependencies = [
    "realesrgan>=0.3.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "opencv-python-headless>=4.5.0",
    "torch>=2.0.0",
    "basicsr>=1.4.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
scaleio = "scaleio.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scaleio*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
